+++ src/imio/smartweb/core/sharing/__init__.py
@@ -1,1 +1,0 @@
-import imio.smartweb.core.sharing.permissions  # NOQA